requirements.txt
venmo-api
python-dotenv
requests
boto3
